data_pipeline/requirements.txt
ccxt
kafka-python
influxdb-client